pyproject.toml
[project]
name = "interconnects-home"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "beautifulsoup4>=4.13.4",
    "pandas>=2.3.1",
    "pypandoc>=1.15",
    "python-frontmatter>=1.1.0",
    "requests>=2.32.5",
]
